+++ Snyk.VisualStudio.Extension.2022/Service/SnykVsSolutionLoadEvents.cs
@@ -108,32 +108,6 @@ namespace Snyk.VisualStudio.Extension.Service
         /// <returns>VSConstants.S_OK.</returns>
         public int OnAfterOpenProject(IVsHierarchy vsHierarchy, int fAdded)
         {
-            var languageClientManager = LanguageClientHelper.LanguageClientManager();
-            if (languageClientManager == null || !(SnykVSPackage.Instance?.IsInitialized ?? false))
-                return VSConstants.S_OK;
-
-            var isFolderTrusted = false;
-            ThreadHelper.JoinableTaskFactory.Run(async () =>
-                isFolderTrusted = await SnykVSPackage.ServiceProvider.TasksService.IsFolderTrustedAsync());
-            if (!isFolderTrusted)
-            {
-                // If folder was not trusted. Stop Language Server if running.
-                languageClientManager.StopServerAsync().FireAndForget();
-                return VSConstants.S_OK;
-            }
-
-            var shouldDownloadCli = SnykVSPackage.ServiceProvider.TasksService.ShouldDownloadCli();
-
-            if (shouldDownloadCli)
-                return VSConstants.S_OK;
-
-            if (languageClientManager.IsReady)
-            {
-                languageClientManager.RestartServerAsync().FireAndForget();
-                return VSConstants.S_OK;
-            }
-
-            languageClientManager.StartServerAsync(true).FireAndForget();
             return VSConstants.S_OK;
         }